requirements.txt
fastapi
uvicorn
streamlit
yfinance
pandas
numpy
openai
python-dotenv
requests
ta
plotly
requests
streamlit-autorefresh